Fix bad memory access in Mersenne twister.
Address sanitizer noticed that the next pointer in MTRand points to
freed memory when used from AtlasPacker. This is because a copy is made
using the default copy constructor. As far as I can tell, there is no
reason to make this copy.

This commit not only removes the needless copy in AtlasPacker, it also
deletes the default copy constructor to prevent this error from being
made elsewhere.