Fix spatial hash that determines colocality.
This fixes a common cause for asserting at the bottom of createColocals
and subsequently crashing (e.g. the GearboxAssy glTF conformance model).

HashMap assumes that if any two keys are "equal", then they map
to the same hash value. The colocality hash was violating this because
it used a epsilon-based equivalency test.

Note that the original Thekla code does not have this issue, it simply
uses the default equivalency test.

On a side note, the equivalency test for Vector3 was not using the same
method that Vector2 was using (Christer Ericson's 2008 post).